Gesture Recognition First Commit
I added a class called Gesture Recognition. To get the mouse gesture,
call the function getGesture from the class GestureRecognizer. the
function throws a GestureNotFoundException if the gesture is not
recognised. There are also 4 gestures implemented at the moment with
their enums in the same class. The function will return an int
representing one of the enums if the gesture is correct.