HMM complete
use the function hmmEvaluate and pass it an array of integers
representing a sequence of directions. the function will return an int
matching one of five enums (gestures). It is important that no two
similar directions follow each other. if the user input returns (NORTH,
NORTH), it should only be passed as (NORTH). The gesture recognition is
very accurate (almost 100%) if the directions are calculated accurately.